Fix classpath error on server (#65)
We encountered an error in the classpath on a specific server, which seems to evaluate to a different version of the commons-lang package. Even when running the same docker image on other servers, we didn't get the same error. We never found the problem, but we solved it by forcing the version of commons-lang.