Hosting resources, certificates and documentation
This PR handles removing limits on hosting resources, as well as documenting how we change certificates in a training environment. It has been used to establish the training environment for the solution. Additionally, the database has been moved to a dev profile, since we will be using SQL as a service for hosting scenarios.

---------

Co-authored-by: Hans Bugge Grathwohl <hbg@globeteam.com>