DDV Integration
DDV Integration

It exists on the same server as FMK, but using different versions of the same schemas. This, combined with the way we make code generation for building the requests, we ran into problems due to overlapping namespaces. Therefore we generate the code, and then change the generated namespace. This is nonstandard use of the code generation, and can cause problems if built incorrectly (and also takes extra time), so we have checked the generated code into the repository, since it does not change often.

---------

Co-authored-by: Hans Bugge Grathwohl <hbg@globeteam.com>