Fetch LMS Data and store it locally, to be accessed when mapping. Also, map the package form code in the CDA.  (#118)
Created object-annotation based parser for LMS data, to facilitate easy extension. Loads LMS Data into a SQLite database to make the data available easily, in an instance of the object. Patched a service into the CDA generator to give access to the data in the static methods.

The integration is reliant on spring, which is not necessary, or really wanted. Feel free to change this.