109: add test data and add dosage unit (#132)
* 109: add many more prescriptions to generator test

And nail them down, so they aren't updated when we run the FmkStorage
refetch.

This allows us to check a more varied list of prescriptions.

* 109: add unit to dosage output

We ended up just using the translated version, because the input is a
free text string. There's an option to improve this later with some
heuristics, but this is enough to go into production.