109: Dosage 1.0 (#137)
* 109: add unstructured dosage text to output

This is the current way that FMK expresses it.

Also handle single dose on a single day cases, change the daily dose to
use hours if more than once per day, and make sure we don't try to
structure doses with a time element - we can't yet.

* 109: add support for periodic intervals over days

* 109: map quantity and consider 'according to need' cases

* 109: handle unbounded cases, where you're free to take them whenever

And also treat AnyDay as a normal Day, since that's only stricter.