New OID implemented in OpenNCP half (#152)
We should not have missed this. We probably need to go over most of our OIDs (also in the international search mask), but I am unsure if this is a good idea to do right in the middle of the test :D...

Also, here is a fix for the repo problem in their git repo. It should be stable even when they fix it.