Service boot test (#272)
This PR makes a small refactoring of the application configuration for country-a-service, and adds a test to our unit test suite that we can load the Spring context, i.e., that we can start the application.

This is to avoid regressions that prevents the application from booting.

The following changes have been made to accommodate this test:

- An option to disable LMS syncing has been added, so that LocalLmsSynchronizer does not block the application startup.  Other options that were considered:  Checking in a database snapshot so the existing logic would prevent a sync.  Connecting to a dummy/mock-FTP server instead of medicinpriser.
- Application configuration has been split into default config  profile which always contains test connection info and no secrets, and the "docker" profile which gets configuration values via environment variables.  Additionally, there is an application config for the new context loading test.
- Configuration validation logic now checks Spring application config values instead of environment variables.