Cleanup from Patient Summary (#278)
This PR is a follow-up on #235. 

Related issue with review comments: #261.

Changes:

Clarification in comments mostly
Removal of ProcessEntry, it might still be relevant, but Jakob has a
good point is that I can't identify why right now, so we are removing
it, and will readd it later if we need it.
Still outstanding:

Refactor of WSDL script. (Will be done later, it works, and I'm not
about to break it before going on vacation and not having tested it
thoroughly enough)
SK/EK clarification: I don't know, we are trying to get a hold of
someone to answer this. Same deal with some of the other clarifications,
I have added todo's for them