Authentication: Exchange OpenNCP SOAP header with IDWS XUA token (#292)
This PR introduces an authentication module which can exchange an
OpenNCP SOAP header for an XCA request (i.e. with HCP and TRC
assertions) to a IDWS XUA token via the bootstrap-to-IDWS service at
SOSI STS.

The entry point for the rest of the application is the class
`dk.sundhedsdatastyrelsen.ncpeh.authentication.AuthenticationService`
which exposes a single method `xcaSoapHeaderToIdwsToken`.

For usage see the integration tests at
`country-a-service/authentication/src/test/java/dk/sundhedsdatastyrelsen/ncpeh/authentication/AuthenticationIT.java`.
To run the integration tests you will need a copy of the keystore with
the test environment certificate.

---------

Co-authored-by: Lillevang <jeppe.lillevang@gmail.com>