Use IDWS tokens for authentification against FMK (#302)
Now, when we call FMK, we use the tokens we get from STS instead
of relying on test identities.

I had to import wss4j in an older version, 1.6.4, because it depends on
santuario which we already use because of seal. This isn't ideal, as that
version has some security flaws, so we should probably either get rid of
seal or fix this some other way.

WSSConfig.init causes loading keys from pkcs12 keystores to fail, so I
switched to jks keystores instead.

I left the non–idws Fmk client in the code, so we can more easily change
back if we need to. Will remove it after more testing.

The configurations for authentication and fmk now need an inputstream
instead of a path, because that makes it easier to use base64 encoded
keystores in integration tests, so we don't have to commit them to git.