FMK role check workaround (#319)
Due to a misunderstanding, the FMK XUA IDWS API has implemented some
checks on the claims in the saml assertion:
- they check that the role of the user is "Pharmacist"
- they check that the organization type is a "Pharmacy".
These are unfortunately not compatible with the EU rules. Also, they
don't work with our (and other countries') test environments.

Until this has been changed at FMK we need to implement a workaround.
This hacky workaround makes string replacements in the SAML assertion.