fix modificator again so effectuations work (#331)
Reverts 28b449d856fe531a47648222586eb4a8917eb376.

Otherwise just adds a comment to a test.

FMK validates the modifier heavily. It must be of type OtherPerson and
ID must be null.

They take the ID from the XUA IDWS header. This is not specified
anywhere yet, but was discovered while both teams worked on this.