always terminate prescription on dispensation
In DK, the pharmacist chooses whether to close the prescription or leave
it open for more dispensations. This is not supported in ePrescription
and eDispensation, so we close all prescriptions. If the patient needs
more medicine, they will have to call their doctor to reopen or
recreate the prescription. But that's better than leaving all
prescriptions open.