Fix integration tests on test2 (#352)
Integration tests on test2 are failing, because the logic for when a
prescription is dispensable didn't match reality. It was hard to test
that on test1, as we were not allowed to automatically create
prescriptions there.

All endpoints have been moved to test2, and "bestilt" is not a blocking
status for a prescription.