add integration test of abort (#355)
There was no great way to do this. Either I had to rewrite most of
submitDispensation in the test, which was a lot of work, or I had to do
something to make submitDispensation fail in the right way. In the end I
opted for a solution that was at least clear in intent, and something
that makes it easy to force aborts for official test rounds as well.