handle prescription dosage differs from medication dosage
Doctors can choose to override the dosage, when they create a
prescription from a drug medication. We now handle this case.

The logic to determine whether the dosage has been overridden depends on
default strings used by FMK, which is not ideal, but there is no other
way currently.

Additionally add short texts to the unstructured texts, as it may happen
that there is only a short text.