update substance unit mapping to all approved ones by LMS (#373)
Lægemiddelstyrelsen has returned with a list of approved mappings of
substance units. The list in the code is now updated to that.

I've double checked with the values found in eHDSIUnit, to make sure we
only send valid units, and there were a few minor discrepancies that
I've fixed, left a comment for, and asked semantics about.

Additionally I've removed a single code (RGD) that we allowed during
tests but is not part of the approved mappings.

Related to #341.