Opt-out service integration (#372)
Integrate the opt-out service in country A service.

To begin with, it is disabled by default in `config/application.yml`.
This is because there is not yet a stable, publicly accessible opt-out
lookup service which can be used for development and automated tests.
Once that is in place, we should enable the integration.