go through all exceptions and ensure messages are OK (#392)
Several exceptions have had their messages changed, so they don't risk
exposing any pii.

Several exceptions have been changed from checked to unchecked, as that
works much better with the functional style we use in the codebase, and
several changes have followed from that.

The Utils class in the FMK api has been removed, it was a duplicate of
the XML utils in base-utils.

Related to #339