certificate documentation and fix (#398)
Updated the readmes to link to a new certificates document in `docs/`.
Added all the certificates I know of that we use and one example of how
to update the certificates.

Additionally re-added the test-tool certificate to the ncp server
truststore.

Additionally changed national-connector to use certificate and key
directly, and not a keystore.